@@ -10,69 +10,68 @@
     "analyze": "ANALYZE=true next build"
   },
   "dependencies": {
-    "@algolia/autocomplete-js": "^1.4.0",
-    "@algolia/autocomplete-plugin-query-suggestions": "^1.4.0",
-    "@algolia/autocomplete-plugin-recent-searches": "^1.4.0",
-    "@algolia/autocomplete-preset-algolia": "^1.4.0",
-    "@algolia/autocomplete-theme-classic": "^1.4.0",
+    "@algolia/autocomplete-js": "^1.4.1",
+    "@algolia/autocomplete-plugin-query-suggestions": "^1.4.1",
+    "@algolia/autocomplete-plugin-recent-searches": "^1.4.1",
+    "@algolia/autocomplete-preset-algolia": "^1.4.1",
+    "@algolia/autocomplete-theme-classic": "^1.4.1",
     "@algolia/react-instantsearch-widget-color-refinement-list": "^1.4.1",
-    "@algolia/react-instantsearch-widget-loadmore-with-progressbar": "^1.3.2",
+    "@algolia/react-instantsearch-widget-loadmore-with-progressbar": "^1.3.5",
     "@algolia/react-instantsearch-widget-size-refinement-list": "^1.0.1",
     "@artsy/fresnel": "^3.0.2",
-    "@material-design-icons/svg": "^0.6.0",
-    "algoliasearch": "^4.10.5",
+    "@material-design-icons/svg": "^0.10.0",
+    "algoliasearch": "^4.11.0",
     "classnames": "^2.3.1",
-    "framer-motion": "^4.1.17",
-    "jotai": "^1.3.6",
-    "next": "^11.1.2",
+    "framer-motion": "^5.0.0",
+    "jotai": "^1.4.2",
+    "next": "^12.0.1",
     "qs": "^6.10.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-fast-compare": "^3.2.0",
-    "react-instantsearch-core": "^6.12.1",
-    "react-instantsearch-dom": "^6.12.1",
-    "search-insights": "^2.0.3"
+    "react-instantsearch-core": "^6.14.0",
+    "react-instantsearch-dom": "^6.14.0",
+    "search-insights": "^2.1.0"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^11.1.2",
+    "@netlify/plugin-nextjs": "^4.0.0-beta.3",
+    "@next/bundle-analyzer": "^12.0.1",
     "@svgr/webpack": "^5.5.0",
-    "@tailwindcss/aspect-ratio": "^0.2.1",
+    "@tailwindcss/aspect-ratio": "^0.3.0",
     "@tweakpane/core": "^1.0.6",
-    "@types/node": "^16.10.1",
+    "@types/node": "^16.11.6",
     "@types/qs": "^6.9.7",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^17.0.33",
+    "@types/react-dom": "^17.0.10",
     "@types/react-instantsearch-dom": "^6.12.0",
-    "@types/tailwindcss": "^2.2.1",
+    "@types/tailwindcss": "^2.2.2",
     "@typescript-eslint/eslint-plugin": "^4.32.0",
     "@typescript-eslint/parser": "^4.32.0",
-    "@welldone-software/why-did-you-render": "^6.2.1",
-    "autoprefixer": "^10.3.6",
+    "autoprefixer": "^10.3.7",
     "babel-eslint": "^10.1.0",
-    "babel-plugin-preval": "^5.0.0",
     "eslint": "^7.32.0",
     "eslint-config-algolia": "^19.0.2",
-    "eslint-config-next": "^11.1.2",
+    "eslint-config-next": "^12.0.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsdoc": "^36.1.0",
+    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-jsdoc": "^37.0.3",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.0",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-unused-imports": "^1.1.5",
     "ifdef-loader": "^2.3.0",
     "next-compose-plugins": "^2.2.1",
-    "next-pwa": "^5.3.1",
-    "postcss": "^8.3.8",
+    "next-pwa": "^5.4.0",
+    "postcss": "^8.3.11",
     "postcss-easy-import": "^3.0.0",
     "prettier": "^2.4.1",
-    "sharp": "^0.29.1",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^22.0.0",
-    "tailwindcss": "^2.2.16",
+    "sharp": "^0.29.2",
+    "stylelint": "^14.0.1",
+    "stylelint-config-standard": "^23.0.0",
+    "tailwindcss": "^2.2.17",
     "tweakpane": "^3.0.5",
-    "typescript": "^4.4.3"
+    "typescript": "^4.4.4"
   }
 }
