@@ -4,10 +4,12 @@ import type { Scope, SetAtom } from 'jotai/core/atom'
 import { useCallback, useEffect, useRef } from 'react'
 import isEqual from 'react-fast-compare'
 
+type ResolveType<T> = T extends Promise<infer V> ? V : T
+
 export function useDeepUpdateAtom<TValue, TUpdate>(
   anAtom: WritableAtom<TValue, TUpdate>,
   scope?: Scope
-): [TValue, SetAtom<TUpdate>] {
+): [ResolveType<TValue>, SetAtom<TUpdate, void>] {
   const [currentAtomValue, _setAtomValue] = useAtom(anAtom, scope)
 
   const currentAtomValueRef = useRef(currentAtomValue)
